fix: cast features to float32 before saving data.pt
The normalization pipeline outputs float64 arrays. PyTorch model
weights are float32. This caused a dtype mismatch error during
forward pass. Cast to float32 at save time.

+++ v2/pipeline/build_v2_dataset.py
@@ -588,7 +588,7 @@ def build_dataset(output_path: str = OUTPUT_PATH):
     oracle_tensors = {}  # no backward compat needed for fresh rebuild
 
     dataset = {
-        'X': torch.from_numpy(X_normalized),
+        'X': torch.from_numpy(X_normalized.astype(np.float32)),
         'feature_names': list(all_feature_names),
 
         'label_trade': torch.from_numpy(label_trade),
